shorts-automation-app/backend/requirements.txt
# FastAPI 및 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터베이스
sqlalchemy==2.0.23
alembic==1.12.1

# API 클라이언트
requests==2.31.0
openai==1.3.7

# Google API (YouTube)
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# 환경 변수
python-dotenv==1.0.0

# 스케줄링
apscheduler==3.10.4

# 유틸리티
python-dateutil==2.8.2
